@@ -5,11 +5,11 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Data extends Model
+class Pasien extends Model
 {
-    protected $fillable = ['namapemilik', 'namahewan', 'jenishewan', 'diagnosa'];
-
-    public function wali(){
-        return $this->belongsTo('App\Wali','wali_id','id');
-    }
+    protected $table = 'datapasiens';
+    protected $fillable = [
+        'nama_hewan','umur_hewan','jenis_hewan','ras_hewan','nama_wali'
+    ];
+    use HasFactory;
 }
