@@ -1,40 +1,41 @@
 @extends('layouts.main')
 
 @section('container')
-
-    <h1>Data Pasien</h1>
-    <a href="/pasien/create" class="btn btn-success">Tambah Data</a>
-    <table class="table table-striped">
+<?php 
+$no=1;
+?>
+<h1>Data Wali</h1>
+    <a href="/pasien/create" class="btn btn-success"> Create </a>
+    <table class="table table-hover">
     <thead>
         <tr>
         <th scope="col">No</th>
-        <th scope="col">Nama Pemilik</th>
         <th scope="col">Nama Hewan</th>
+        <th scope="col">Umur Hewan</th>
         <th scope="col">Jenis Hewan</th>
-        <th scope="col">Diagnosa</th>
-        <th scope="col"></th>
+        <th scope="col">Ras Hewan</th>
+        <th scope="col">Nama Wali</th>
+        <th scope="col">Aksi</th>
+        <th></th>
         </tr>
     </thead>
     <tbody>
         @foreach ($pasiens as $pasien)
-        <tr>
-        <td>{{$loop->iteration}}</td>
-        <td>{{$pasien->namapemilik}}</td>
-        <td>{{$pasien->namahewan}}</td>
-        <td>{{$pasien->jenishewan}}</td>
-        <td>{{$pasien->diagnosa}}</td>
-        <td>
-            <a href="{{route('pasien.edit', $pasien->id)}}" class="btn btn-warning">Ubah</a>
-        </td>
-        <form action="{{route('pasien.destroy', $pasien->id)}}" method="POST">
-                    @csrf
-                    @method('DELETE')
-                <input type="submit" class="btn btn-danger" value="Delete">
-            </form>
-        </tr>
+            <tr>
+            <td>{{$no++}}</td>
+            <td>{{$pasien->nama_hewan}}</td>
+            <td>{{$pasien->umur_hewan}}</td>
+            <td>{{$pasien->jenis_hewan}}</td>
+            <td>{{$pasien->ras_hewan}}</td>
+            <td>{{$pasien->nama_wali}}</td>
+            <td>
+                <a href="/pasien/{{$pasien->id}}/edit" class="btn btn-primary">Edit</a>
+                <a href="/pasien/delete" class="btn btn-danger"><span data-feather="edit">Delete</span></a>
+            </td>
+            </tr>
         @endforeach
-    </tbody>
-    </table>
-    @stop
+
+  </tbody>
+</table>
     <a href="/home">Back to Home</a>
 @endsection
